+++ Demo/src/main/java/com/paypal/android/api/model/PayPalSetupToken.kt
@@ -1,5 +1,8 @@
 package com.paypal.android.api.model
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class PayPalSetupToken(
     val id: String,
     val customerId: String,