+++ PayPalWebPayments/src/main/java/com/paypal/android/paypalwebpayments/PayPalWebStartCallback.kt
@@ -1,0 +1,12 @@
+package com.paypal.android.paypalwebpayments
+
+import androidx.annotation.MainThread
+
+fun interface PayPalWebStartCallback {
+
+    /**
+     * Called when the result of a PayPal web launch is available.
+     */
+    @MainThread
+    fun onPayPalWebStartResult(result: PayPalPresentAuthChallengeResult)
+}